+++ packages/fabricator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pact-toolbox/fabricator
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pact-toolbox/utils@0.0.6
+
 ## 0.0.1
 
 ### Patch Changes