+++ packages/installer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pact-toolbox/installer
 
+## 0.0.7
+
+### Patch Changes
+
+- Improve installer version detection
+- Updated dependencies
+  - @pact-toolbox/utils@0.0.6
+
 ## 0.0.6
 
 ### Patch Changes