+++ packages/network/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pact-toolbox/network
 
+## 0.0.9
+
+### Patch Changes
+
+- Improve installer version detection
+- Updated dependencies
+  - @pact-toolbox/utils@0.0.6
+  - @pact-toolbox/prelude@0.0.9
+  - @pact-toolbox/proxy@0.0.6
+
 ## 0.0.8
 
 ### Patch Changes