+++ packages/unplugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pact-toolbox/unplugin
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @pact-toolbox/network@0.0.9
+  - @pact-toolbox/utils@0.0.6
+
 ## 0.0.9
 
 ### Patch Changes